result_saver/requirements.txt
confluent-kafka>=2.0.0
psycopg2-binary>=2.9.0